requirements.txt
amqp==5.3.1
asgiref==3.9.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.1
cachetools==5.5.2
celery==5.5.3
certifi==2025.8.3
cffi==1.17.1
channels==4.3.1
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
cron_descriptor==2.0.6
cryptography==45.0.7
daphne==4.2.1
Django==5.2.6
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-redis==6.0.0
django-timezone-field==7.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
geographiclib==2.1
geopy==2.4.1
google-auth==2.40.3
google-auth-oauthlib==1.2.2
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
kombu==5.5.4
oauthlib==3.3.1
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.23.0
PyJWT==2.10.1
pyOpenSSL==25.1.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
redis==6.4.0
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
zope.interface==7.2
orjson
firebase-admin==6.6.0
